+++ test/aliases.js
@@ -1,5 +1,5 @@
 function emoji_span(codepoint){
-	return '<span class="emoji emoji-sizer" style="background-image:url(/'+codepoint+'.png)"></span>';
+	return '<span class="emoji emoji-sizer" style="background-image:url(/'+codepoint+'.png)" data-codepoints="'+codepoint+'"></span>';
 }
 
 describe("Aliases", function(){