+++ test/modes.spec.js
@@ -20,7 +20,7 @@ describe("Correctly supports different modes", function(){
 		emoji.use_sheet = true;
 		emoji.use_css_imgs = false;
 
-		expect(emoji.replace_colons(':cloud:')).toBe('<span class="emoji-outer emoji-sizer"><span class="emoji-inner" style="background: url(/sheet.png);background-position:2.5% 2.5%;background-size:4100%"></span></span>');
+		expect(emoji.replace_colons(':cloud:')).toBe('<span class="emoji-outer emoji-sizer"><span class="emoji-inner" style="background: url(/sheet.png);background-position:2.5% 2.5%;background-size:4100%" data-codepoints="2601"></span></span>');
 	});
 
 	it("Uses CSS classes with CSS background-sizing support", function(){
@@ -29,7 +29,7 @@ describe("Correctly supports different modes", function(){
 		emoji.use_sheet = false;
 		emoji.use_css_imgs = true;
 
-		expect(emoji.replace_colons(':cloud:')).toBe('<span class="emoji emoji-2601"></span>');
+		expect(emoji.replace_colons(':cloud:')).toBe('<span class="emoji emoji-2601" data-codepoints="2601"></span>');
 	});
 
 	it("Uses individual images with CSS background-sizing support", function(){
@@ -38,7 +38,7 @@ describe("Correctly supports different modes", function(){
 		emoji.use_sheet = false;
 		emoji.use_css_imgs = false;
 
-		expect(emoji.replace_colons(':cloud:')).toBe('<span class="emoji emoji-sizer" style="background-image:url(/2601.png)"></span>');
+		expect(emoji.replace_colons(':cloud:')).toBe('<span class="emoji emoji-sizer" style="background-image:url(/2601.png)" data-codepoints="2601"></span>');
 	});
 
 	it("Uses images without CSS background-sizing support", function(){
@@ -47,7 +47,7 @@ describe("Correctly supports different modes", function(){
 		emoji.use_sheet = false;
 		emoji.use_css_imgs = false;
 
-		expect(emoji.replace_colons(':cloud:')).toBe('<img src="/2601.png" class="emoji" />');
+		expect(emoji.replace_colons(':cloud:')).toBe('<img src="/2601.png" class="emoji" data-codepoints="2601" />');
 	});
 
 });