+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * Use native emoji in Chrome and Firefox on OS X (thanks @CharlieHess)
 * Handle being in a hidden iframe on Firefox - https://github.com/iamcal/js-emoji/issues/73
+* Added `data-codepoints` attribute to all output modes
 
 ## 3.0.2 - 2016-06-17